# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dar Al Maghrib Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/me/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # CATEGORIES
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{menuItemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
...
  /admin/orders:
    get:
      operationId: adminGetOrders
      tags: [admin]
      summary: Get all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/reviews/{id}:
    delete:
      operationId: adminDeleteReview
      tags: [admin]
      summary: Delete review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, admin]
        createdAt:
          type: string
          format: date-time

    # CATEGORIES
    Category:
      type: object
      required: [id, name, nameAr, nameFr, icon, sortOrder, active]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        icon:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        active:
          type: boolean
        itemCount:
          type: integer
...
    MenuItem:
      type: object
      required: [id, name, nameAr, nameFr, description, descriptionAr, descriptionFr, price, categoryId, available, spiceLevel, preparationTime, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        available:
          type: boolean
        spiceLevel:
          type: string
          enum: [none, mild, medium, hot, very_hot]
        preparationTime:
          type: integer
          description: In minutes
        calories:
          type: ["integer", "null"]
        ingredients:
          type: string
        isPopular:
          type: boolean
        isFeatured:
          type: boolean
        isTodaySpecial:
          type: boolean
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    CartItem:
      type: object
      required: [menuItemId, quantity, menuItem]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        menuItem:
          $ref: "#/components/schemas/MenuItem"
    Cart:
      type: object
      required: [items, subtotal, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        total:
          type: number
    CartItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # ORDERS
    OrderItem:
      type: object
      required: [menuItemId, quantity, unitPrice, totalPrice, menuItemName]
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemImageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    Order:
      type: object
      required: [id, userId, status, orderType, paymentMethod, subtotal, total, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, preparing, out_for_delivery, delivered, cancelled]
        orderType:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
          enum: [online, cash_on_delivery]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        deliveryNotes:
          type: ["string", "null"]
        couponCode:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        total:
          type: number
        estimatedDeliveryTime:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, accepted, preparing, out_for_delivery, delivered, cancelled]

    # REVIEWS
...