# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dar Al Maghrib Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItem } from './menuItem';

export interface CartItem {
  menuItemId: number;
  quantity: number;
  menuItem: MenuItem;
}