# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dar Al Maghrib Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderOrderType } from './orderOrderType';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  status: OrderStatus;
  orderType: OrderOrderType;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  customerName?: string;
  customerPhone?: string;
  customerEmail?: string;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryNotes?: string | null;
  /** @nullable */
  couponCode?: string | null;
  subtotal: number;
  deliveryFee?: number;
  discount?: number;
  total: number;
  /** @nullable */
  estimatedDeliveryTime?: number | null;
  createdAt: Date;
  updatedAt?: Date;
  items: OrderItem[];
}